mod main-test

+++ ft_strnstr.c
@@ -14,8 +14,8 @@
 
 char    *ft_strnstr(const char *s1, const char *s2, size_t len)
 {
-    size_t pos;
-    int i;
+    size_t  pos;
+    size_t  i;
 
     pos = 0;
     if (!*s2)
